format string expects integers. cast calculated values to ints. reviewed by jcausey.
git-svn-id: http://dev.zenoss.com/svnint/trunk/core@33839 f5df01ce-0320-0410-9ed9-ddcd13c732cc

+++ Products/ZenModel/Device.py
@@ -1001,10 +1001,10 @@ class Device(ManagedEntity, Commandable, Lockable, MaintenanceWindowable,
         elif ut == 0:
             return "0d:0h:0m:0s"
         ut = float(ut)/100.
-        days = ut/86400
-        hour = (ut%86400)/3600
-        mins = (ut%3600)/60
-        secs = ut%60
+        days = int(ut/86400)
+        hour = int((ut%86400)/3600)
+        mins = int((ut%3600)/60)
+        secs = int(ut%60)
         return "%02dd:%02dh:%02dm:%02ds" % (
             days, hour, mins, secs)
